refactor: web api to es6
