+++ src/utils/user.js
@@ -7,7 +7,8 @@ import {stringToMD5} from './string';
 export function generateGravatarUrl(email?: string): string {
   if (typeof email === 'string') {
     email = email.trim().toLocaleLowerCase();
-    let emailMD5 = stringToMD5(email);
+    const emailMD5 = stringToMD5(email);
+
     return `https://www.gravatar.com/avatar/${emailMD5}`;
   } else {
     return 'https://www.gravatar.com/avatar/00000000000000000000000000000000?d=mm';
